requirements/runtime.txt
fastapi
uvicorn
pydantic
PyYAML
python-dotenv
jinja2
click
prompt_toolkit
croniter
watchdog
fastjsonschema
psutil
screeninfo
pywin32
opencv-python
numpy
paddleocr
